third_party/WebKit/LayoutTests/external/wpt/html/dom/dynamic-markup-insertion/opening-the-input-stream/document.open-03-expected.txt

This is a testharness.js-based test.
FAIL document.open and singleton replacement assert_equals: setTimeout expected false but got true
Harness: the test ran to completion.
